methodlistAuto(Consense)	R Documentation

methodlistAuto

Description

Given an input kinput (number of clusters), generate a methodlist for use by link{clusterList}.

Usage

methodlistAuto(methodlistAuto(kinput)

Arguments

kinput 	 A desired geometry (number of clusters) k, greater than 1 and less than n (number of genes or samples).

Details

Extending this code is potentially the easiest way to add a new method for benchmarking. Details to follow.

Value

methodlist 	 A list of methods and associated parameters passed on to link{clusterList}.

Author(s)

Ted Laderas (laderast@ohsu.edu)

See Also

getClusterInfo

[Package Consense version 0.1 Index]